requirements.txt
Django==1.11.7
django-ranged-response==0.2.0
django-simple-captcha==0.5.7
Pillow==5.1.0
pytz==2018.4
six==1.11.0
